Data deletion DONE

+++ src/lib/core/utils/file_handler.py
@@ -76,32 +76,38 @@ def multi_file_open(path):
 
 def delete_file_directory(path: Union[Path, str]):
 
-    # convert path string into path-like object
-    path = Path(path) if isinstance(path, str) else path
-    log_info(path)
+    if path:
+        # convert path string into path-like object
+        path = Path(path) if isinstance(path, str) else path
+        log_info(path)
 
-    if path.is_dir():
+        if path.is_dir():
 
-        # remove all contents if path is a folder directory
-        try:
-            shutil.rmtree(path)
-            log_info(f"Successfully deleted {path.name} folder")
-        except Exception as e:
-            error_msg = f"{e}: Failed to remove directory"
-    elif path.is_file():
+            # remove all contents if path is a folder directory
+            try:
+                shutil.rmtree(path)
+                log_info(f"Successfully deleted {path.name} folder")
+                return True
+            except Exception as e:
+                error_msg = f"{e}: Failed to remove directory"
+                return False
+        elif path.is_file():
 
-        # remove file if path is filepath
+            # remove file if path is filepath
 
-        try:
-            path.unlink()
-            log_info(f"Successfully deleted {path.name}")
+            try:
+                path.unlink()
+                log_info(f"Successfully deleted {path.name}")
+                return True
 
-        except FileNotFoundError as e:
-            error_msg = f"{e}: {path.name} does not exists"
-            log_error(error_msg)
-    else:
-        not_path_error_msg = f"Invalid path given......"
-        log_error(not_path_error_msg)
+            except FileNotFoundError as e:
+                error_msg = f"{e}: {path.name} does not exists"
+                log_error(error_msg)
+                return False
+        else:
+            not_path_error_msg = f"Invalid path given......"
+            log_error(not_path_error_msg)
+            return False
 
 
 def file_search(path=str(Path.home())):
